server/requirements.txt
aiofiles    
aiohappyeyeballs
aiosignal
aiosqlite
annotated-types
anyio
APScheduler
async-timeout
attrs
backoff
backports-datetime-fromisoformat
bcrypt
beautifulsoup4
Brotli
build
cachetools
certifi
cffi
chardet
charset-normalizer
click
click-log
cobble
colorama
coloredlogs
cryptography
cssselect
deprecation
dill
distro
durationpy
ecdsa
exceptiongroup
fake-http-header
fake-useragent
fastapi
filelock
firecrawl
flatbuffers
frozenlist
fsspec
google-auth
googleapis-common-protos
gotrue
greenlet
grpcio
h11
h2
hpack
httpcore
httptools
httpx
huggingface-hub
humanfriendly
humanize
hyperframe
idna
importlib_metadata
importlib_resources
Jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema-specifications
lark
lxml
mammoth
markdown-it-py
MarkupSafe
marshmallow
mdurl
mmh3
mpmath
multidict
multiprocess
nest-asyncio
networkx
nltk
numpy
oauthlib
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-proto
orjson
outcome
overrides
packaging
pandas
pillow
playwright
postgrest
posthog
propcache
protobuf
psutil
psycopg2-binary
pyarrow
pyasn1
pyasn1_modules
pybase64
pycparser
pydantic
pydantic_core
pyee
pygitguardian
Pygments
PyJWT
PyPDF2
pyperclip
PyPika
pyproject_hooks
pyreadline3
PySocks
python-dateutil
python-docx
python-dotenv
python-jose
python-multipart
pytz
PyYAML
rank-bm25
realtime
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
rtree
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
shapely
shellingham
six
sniffio
snowballstemmer
sortedcontainers
soupsieve
SQLAlchemy
starlette
storage3
StrEnum
supabase
supafunc
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
tomli
torch
tqdm
transformers
trimesh
trio
typing-inspection
typing_extensions
tzdata
tzlocal
unicorn
urllib3
uvicorn
webdriver-manager
websocket-client
websockets
wsproto
xxhash
yarl
zipp
zstandard
selenium
undetected_chromedriver